Changes - Handle expiry dates - Split LobbyCommand logic - Add status command - Fix and improve a lot of things
- Regi will not handle the EXPIRED status
- Regi will now handle lobbies expiring
        - Added "defaultExpiry" to config; a format whose map type does not have an override will use this expiry time
- Regi will now handle Lobby names
- Fixed the "Unknown Channel" error appearing due to not checking for its existence before deleting
- Deleted the fake error message when a channel deletion fails
- Split LobbyCommand's code into files for better readability and commodity
- Added /lobby status sub-command
       - Can only be ran on the Administrator channel
       - Will list all active lobbies currently, and if parsed a Lobby ID, will list all details of it
- Improved Interaction replies to be all deferred before replying
       - Also fixes the "Unknown Interaction" error
- LobbyService will now monitor every 7 seconds for expired lobbies
- Implemented a config file system for each format & map type
      - Every map type can have it's own expiry time (for Lobby) and config file

+++ src/discord.service.ts
@@ -180,9 +180,16 @@ export class DiscordService {
       if (id !== undefined) {
         try {
           // Get the channel with that ID and delete it
-          const channel = await this.bot.channels.fetch(id);
+          try {
+            const channel = await this.bot.channels.fetch(id);
 
-          await channel.delete();
+            if (channel) await channel.delete();
+          } catch (e) {
+            this.logger.error(
+              `Could not find channel with ID ${id}. Skipping...`,
+            );
+            continue;
+          }
         } catch (e) {
           this.logger.error(
             `Tried to delete channel ${id} from Lobby ${lobby.id}: ${e}.`,
